refactor(AmountInput):remove barrel files
