--- libs/ui-react/src/lib/Components/AmountInput/useAmountInputValue/index.ts
@@ -1,1 +1,0 @@
-export { useAmountInputValue } from './useAmountInputValue';
