--- libs/ui-react/src/lib/Components/AmountInput/useAutoWidthInput/index.ts
@@ -1,1 +1,0 @@
-export { useAutoWidthInput } from './useAutoWidthInput';
