--- libs/ui-rnative/src/lib/Components/AmountInput/useAmountInputAnimations/index.ts
@@ -1,1 +1,0 @@
-export { useAmountInputAnimations } from './useAmountInputAnimations';
