--- libs/ui-rnative/src/lib/Components/AmountInput/useAmountInputFormatting/index.ts
@@ -1,1 +1,0 @@
-export { useAmountInputFormatting } from './useAmountInputFormatting';
